College-Bot/requirements.txt
flask
google-generativeai
google-cloud-aiplatform